requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
conda-pack==0.6.0
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
einops==0.3.2
facenet-pytorch==2.5.2
google-auth==1.33.0
google-auth-oauthlib==0.4.1
grad-cam==1.3.5
grpcio==1.36.1
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.8.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.60
pandas==1.1.5
Pillow==8.4.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorboardX==2.2
timm==0.4.12
torch==1.10.0+cu113
torch-tb-profiler==0.3.1
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
ttach==0.0.3
typing_extensions==4.0.0
urllib3==1.26.7
vit-pytorch==0.24.2
Werkzeug==2.0.2
yarl==1.6.3
zipp==3.6.0
